refactor: satisfy dependency inversion a bit more
