+++ packages/react-native-reanimated/src/hook/commonTypes.ts
@@ -5,6 +5,7 @@ import type {
   ShadowNodeWrapper,
   SharedValue,
   WorkletFunction,
+  AnimatedStyle,
 } from '../commonTypes';
 import type {
   ImageStyle,
@@ -14,7 +15,6 @@ import type {
   NativeScrollEvent,
 } from 'react-native';
 import type { ViewDescriptorsSet } from '../ViewDescriptorsSet';
-import type { AnimatedStyle } from '../helperTypes';
 import type { ReanimatedHTMLElement } from '../js-reanimated';
 
 export type DependencyList = Array<unknown> | undefined;
