+++ packages/react-native-reanimated/src/index.ts
@@ -93,7 +93,6 @@ export {
   useInterpolateConfig,
 } from './interpolateColor';
 export type {
-  EasingFunction,
   EasingFn,
   EasingFunctionFactory,
   EasingFactoryFn,
@@ -245,6 +244,12 @@ export type {
   AnimatedKeyboardInfo,
   AnimatedKeyboardOptions,
   MeasuredDimensions,
+  EasingFunction,
+  AnimatedTransform,
+  TransformArrayItem,
+  AnimateStyle,
+  AnimatedStyle,
+  StylesOrDefault,
 } from './commonTypes';
 export {
   SensorType,
@@ -272,13 +277,8 @@ export type {
   AdaptTransforms,
   AnimateProps,
   AnimatedProps,
-  AnimatedTransform,
   TransformStyleTypes,
-  TransformArrayItem,
-  AnimateStyle,
-  AnimatedStyle,
   AnimatedStyleProp,
-  StylesOrDefault,
 } from './helperTypes';
 export type { AnimatedScrollViewProps } from './component/ScrollView';
 export type { FlatListPropsWithLayout } from './component/FlatList';
