+++ packages/react-native-reanimated/src/js-reanimated/index.ts
@@ -1,7 +1,6 @@
 'use strict';
 import JSReanimated from './JSReanimated';
-import type { StyleProps } from '../commonTypes';
-import type { AnimatedStyle } from '../helperTypes';
+import type { StyleProps, AnimatedStyle } from '../commonTypes';
 import {
   createReactDOMStyle,
   createTransformValue,
