+++ packages/react-native-reanimated/src/layoutReanimation/animationBuilder/commonTypes.ts
@@ -1,7 +1,10 @@
 'use strict';
-import type { TransformArrayItem } from '../../helperTypes';
-import type { EasingFunction } from '../../Easing';
-import type { ShareableRef, StyleProps } from '../../commonTypes';
+import type {
+  ShareableRef,
+  StyleProps,
+  TransformArrayItem,
+  EasingFunction,
+} from '../../commonTypes';
 
 export type LayoutAnimationsOptions =
   | 'originX'
